🦕 [BE: feat] Pagination

See merge request s08-final/S08P31E203!54

+++ ourkitty/src/main/kotlin/com/nyang/ourkitty/common/dto/ResultDto.kt
@@ -2,5 +2,5 @@ package com.nyang.ourkitty.common.dto
 
 data class ResultDto<T>(
     val data: T,
-    val totalCount: Int = 1,
+    val totalCount: Long = 1L,
 )
